pyproject.toml
# Copyright (c) Ansible Project
# GNU General Public License v3.0+ (see LICENSES/GPL-3.0-or-later.txt or https://www.gnu.org/licenses/gpl-3.0.txt)
# SPDX-License-Identifier: GPL-3.0-or-later

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "antsibull-nox"
dynamic = ["version"]
description = "Changelog tool for Ansible-core and Ansible collections"
readme = "README.md"
requires-python = ">= 3.9.0"
license = "GPL-3.0-or-later"
license-files.globs = ["LICENSES/*.txt"]
authors = [
  { name = "Felix Fontein", email = "felix@fontein.de" },
  { name = "Toshio Kuratomi", email = "a.badger@gmail.com" },
  { name = "Matt Clay", email = "matt@mystile.com" },
]
maintainers = [
  { name = "Felix Fontein", email = "felix@fontein.de" },
  { name = "Maxwell G", email = "maxwell@gtmx.me" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Ansible",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "antsibull-fileutils >= 1.4.0, < 2.0.0",
    "nox >= 2025.2.9",
    "packaging",
    "pydantic ~= 2.0",
    "pyyaml",
    "semantic_version",
    "tomli; python_version<'3.11'",
]

[project.urls]
"Documentation" = "https://ansible.readthedocs.io/projects/antsibull-nox/"
"Source code" = "https://github.com/ansible-community/antsibull-nox/"
"Code of Conduct" = "https://docs.ansible.com/ansible/latest/community/code_of_conduct.html"
"Bug tracker" = "https://github.com/ansible-community/antsibull-nox/issues"
"Changelog" = "https://github.com/ansible-community/antsibull-nox/tree/main/CHANGELOG.md"

[project.scripts]
antsibull-nox = "antsibull_nox.cli:main"


[project.optional-dependencies]
#
# Development. Not user facing.
#

codeqa = [
    "antsibull-changelog",
    "flake8 >= 3.8.0",
    "pylint",
    "reuse",
]
coverage = [
    "coverage[toml]",
]
formatters = [
    "black>=24",
    "isort",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-error-for-skips",
]
typing = [
    "mypy",
    "antsibull-docutils >= 1.2.0",
    "types-docutils",
    "types-PyYAML",
    "types-jinja2",
]
dev = [
    # Used by nox sessions
    "antsibull-nox[codeqa]",
    "antsibull-nox[coverage]",
    "antsibull-nox[formatters]",
    "antsibull-nox[test]",
    "antsibull-nox[typing]",
    # misc
    "nox",
]

[tool.hatch.version]
path = "src/antsibull_nox/__init__.py"

[tool.isort]
profile = "black"

[tool.coverage.paths]
source = [
    "src",
    "*/site-packages",
]

[[tool.mypy.overrides]]
module = [
    "ansible.module_utils.basic",
    "pytest",
    "semantic_version",
    "yamllint",
    "yamllint.cli",
    "yamllint.config",
    "yamllint.linter",
]
ignore_missing_imports = true
